+++ src/AppProvider.js
@@ -5,11 +5,14 @@ import { Provider } from 'react-redux'
 import { getUserStorageData } from 'shared/utilities/storageUtils'
 import App from 'App'
 import ReactGA from 'react-ga'
-ReactGA.initialize('UA-137162338-1', {
-  debug: process.env.NODE_ENV === 'development',
-  testMode: process.env.NODE_ENV === 'development',
-  anonymizeIp: true, // Removes last 3 digits from IP
-})
+import * as c from 'shared/constants'
+if (c.ENABLE_GOOGLE_ANALYTICS) {
+  ReactGA.initialize('UA-137162338-1', {
+    debug: process.env.NODE_ENV === 'development',
+    testMode: process.env.NODE_ENV === 'development',
+    anonymizeIp: true, // Removes last 3 digits from IP
+  })
+}
 ReactGA.pageview(window.location.pathname + window.location.search)
 
 import 'react-toastify/dist/ReactToastify.css'